Refactor directive module to folder
